run addon helpers on componentDidUpdate
This fixes an issue where changing tabs broke pages with asides because
the tab switch re-wrote the DOM.
TODO: work on a better way to get the postProcess function to nexted children as it is not needed for
container components. There is a react pattern for this, just forget how to do it.